moved files and readme edit
