+++ frontend/src/comp/Btn.jsx
@@ -1,0 +1,7 @@
+export const Button = () => {
+  return (
+    <button>
+      Button
+    </button>
+  );
+}
