+++ frontend/src/comp/user/LogoutBtn.jsx
@@ -1,5 +1,5 @@
 import { useNavigate } from 'react-router-dom';
-import useUserStore from '../stores/useUserStore';
+import useUserStore from '../../stores/useUserStore';
 
 export const Logout = () => {
   const clearUser = useUserStore((state) => state.clearUser);
